removed redundant inventoryListAdmin component
